Only clear cache if something was deleted.
svn commit r35876

+++ Site/admin/components/Sidebar/Delete.php
@@ -31,7 +31,7 @@ class SiteSidebarDelete extends AdminDBDelete
 
 		$num = SwatDB::exec($this->app->db, $sql);
 
-		if (isset($this->app->memcache)) {
+		if ($num > 0 && isset($this->app->memcache)) {
 			$this->app->memcache->delete('gadget_instances');
 		}
